Small fix for setting the DateTimeFormat of an NZazu date-Field
When the Field is created, it automatically sets the DateTimeFormat that is found inside of the provided Settings, and sets the wanted enum to "Custom".

But with the Function SetProperty() --> GetConvertedValue() inside of the class PropertyExtensions the Value will be checked again and will fail most of the time, and leave the users confused as to why an error is Traced.
This change will fix that