Provide Information for a failed SetProperty Attempt
Before there were no information about which Field is failing to set its Properties. The Exception that is caught inside the PropertyExtension will now be thrown again as an ArgumentException (because the caught Exception is always a normal Exception... which is not that optimal) and will be caught where the settings are actually applied. Here you'll have the ability to trace which field (with the key of that field) is affected, so searching for an error will be easier.