Restrict "AutoCompleteField" to only on "AutoCompleteManager"
In some rare cases it may happen, that the "Loaded" Event of the "TextBox" will be called twice. Because an "AutoCompleteManager" will be added to the "TextBox" each time the "Loaded" Event is fired, it may be added multiple times, resulting in having more than one "AutoCompleteManager". Which results in displaying multiple popups, when the User writes something.